fix: Refactor ChatRoom component
